+++ tests/testthat/test-importData-queryData.R
@@ -1,4 +1,4 @@
-testthat::test_that("Test module queryDataServer", {
+testthat::test_that("Test queryDataServer", {
   testMergeList <-
     readRDS(testthat::test_path("test-importData-mergeData_data.rds"))
 
@@ -51,12 +51,12 @@ testthat::test_that("Test module queryDataServer", {
   dbWriteTable(testdb, "table1", testTable1)
   dbWriteTable(testdb, "table2", testTable2)
 
-  expect_equal(dbListTables(testdb), c("table1", "table2"))
+  testthat::expect_equal(dbListTables(testdb), c("table1", "table2"))
 
   testQuery <-
     "SELECT t1.`Human.Entry.ID`, t1.`Age.Category`, t1.`Site.Name` FROM table1 AS t1 LEFT JOIN table2 as t2 ON t1.`Age.Category` = t2.`Age.Category` AND t1.`Site.Name` = t2.`Site.Name`;"
 
-  expect_equal(
+  testthat::expect_equal(
     RSQLite::dbGetQuery(testdb, testQuery),
     structure(
       list(
@@ -74,7 +74,7 @@ testthat::test_that("Test module queryDataServer", {
   testQueryFailure <-
     "SELECT `Human.Entry.ID`, `Age.Category`, `Site.Name` FROM table1 LEFT JOIN table2 ON `Age.Category` = `Age.Category` AND `Site.Name` = `Site.Name`;"
 
-  expect_warning(tryCatch({
+  testthat::expect_warning(tryCatch({
     RSQLite::dbGetQuery(testdb, testQueryFailure)
     #stop("test error")
     #warning("test warning")
@@ -90,4 +90,43 @@ testthat::test_that("Test module queryDataServer", {
     return(NULL)
   },
   finally = NULL))
+})
+
+
+testthat::test_that("Test gptServer", {
+  testthat::expect_error(validateKey(
+    testthat::test_path("test-importData_gpt3_invalidKeyFormat.txt")
+  ))
+  testthat::expect_error(validateKey(
+    testthat::test_path("test-importData_gpt3_validKeyFormat.txt")
+  ), regexp = NA)
+
+  shiny::testServer(gptServer,
+                    args = list(autoCompleteList = reactive(c("testA", "testB"))),
+                    {
+                      # Arrange
+                      print("test gptServer: no auth, empty prompt")
+                      # Act
+                      session$setInputs(
+                        apiKey =
+                          structure(
+                            list(
+                              name = "test-importData_gpt3_validKeyFormat.txt",
+                              size = 10L,
+                              type = "text/plain",
+                              datapath = testthat::test_path("test-importData_gpt3_validKeyFormat.txt")
+                            ),
+                            class = "data.frame",
+                            row.names = c(NA, -1L)
+                          ),
+                        temperature = 0.1,
+                        maxTokens = 100,
+                        n = 1,
+                        gptPrompt = "",
+                        applyPrompt = 1
+                      )
+
+                      testthat::expect_false(validConnection())
+                      testthat::expect_null(session$returned())
+                    })
 })