+++ NEWS.md
@@ -1,5 +1,21 @@
 # DataTools
 
+## Version 23.02.8
+
+### New Features
+- _Import Data_ module in the tab _Query with SQL_: option to use GPT3 for creation of SQL queries
+
+## Version 23.02.7
+
+### New Features
+- New function to catch and forward errors and warnings to the UI of the app. If a warning is 
+triggered the result of the evaluated expression is kept and returned.
+
+## Version 23.02.6
+
+### Updates
+- factors are converted to character in cutAllLongStrings (#25)
+
 ## Version 23.02.5
 
 ### Bug Fixes